Move get_version to util.
